+++ CHANGELOG.md
@@ -6,6 +6,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 #### Added
+#### Fixed
+#### Removed
+
+
+
+## [1.3.0] - 2017-06-07
+#### Added
 - **[CORE]** Support for System Applications
 - **[CORE]** Global variable `HIDE_SYSTEM_APPS`: if set to `True`, only 3rd party apps will be shown
 - **[MODULE]** `SignerIdentity` included in `binary/info/metadata`
@@ -14,8 +21,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 #### Fixed
 - **[MODULE]** Case sensitivity issues in `device/dependency_installer` _[from @tghosth]_
-
-#### Removed